// src/BindingChaos.Web/src/api/models/SignalViewModel.ts
export interface SignalViewModel {
    /**
     * The unique identifier of the signal.
     * @type {string}
     * @memberof SignalViewModel
     */
    id?: string;
    /**
     * The title of the signal.
     * @type {string}
     * @memberof SignalViewModel
     */
    title?: string;
    /**
     * The description of the signal.
     * @type {string}
     * @memberof SignalViewModel
     */
    description?: string;
    /**
     * The tags associated with the signal.
     * @type {Array<string>}
     * @memberof SignalViewModel
     */
    tags?: Array<string>;
    /**
     * The pseudonym of the author.
     * @type {string}
     * @memberof SignalViewModel
     */
    authorPseudonym?: string;
    /**
     * The creation timestamp in ISO 8601 format.
     * @type {string}
     * @memberof SignalViewModel
     */
    createdAt?: string;
    /**
     * The number of amplifications for this signal.
     * @type {number}
     * @memberof SignalViewModel
     */
    amplificationCount?: number;
    /**
     * The number of comments for this signal.
     * @type {number}
     * @memberof SignalViewModel
     */
    commentCount?: number;
    /**
     * Indicates whether the current user has amplified this signal.
     * @type {boolean}
     * @memberof SignalViewModel
     */
    isAmplifiedByCurrentUser?: boolean;
    /**
     * Indicates whether the current user is the originator of this signal.
     * @type {boolean}
     * @memberof SignalViewModel
     */
    isOriginator?: boolean;
    /**
     * URL to the first attachment thumbnail (if any images exist).
     * This provides a visual preview for the signal in the list view.
     * @type {string}
     * @memberof SignalViewModel
     */
    firstAttachmentThumbnail?: string | null;
    /**
     * Total number of attachments associated with this signal.
     * @type {number}
     * @memberof SignalViewModel
     */
    attachmentCount?: number;
}